Move strings to their respective screens
