+++ src/app/screens/App/screens/Home/strings.js
@@ -4,6 +4,3 @@ export const AUTHOR = 'Autor';
 export const SELECT_FILTER = 'Seleccionar filtro';
 export const SEARCH = 'Buscar...';
 export const SEARCH_BUTTON_ALT_TEXT = 'Buscar';
-export const LOGO_ALT_TEXT = 'Logo de WBooks';
-export const LOGIN_WITH_GOOGLE = 'INGRESAR CON GOOGLE';
-export const SIGN_UP = 'REGISTRARSE';
