+++ src/app/screens/App/screens/Landing/strings.js
@@ -1,0 +1,3 @@
+export const LOGO_ALT_TEXT = 'Logo de WBooks';
+export const LOGIN_WITH_GOOGLE = 'INGRESAR CON GOOGLE';
+export const SIGN_UP = 'REGISTRARSE';
