+++ src/app/screens/App/screens/Login/strings.js
@@ -1,6 +1,4 @@
 export const LOGIN = 'Login';
-export const EMAIL = 'Email';
-export const PASSWORD = 'Contraseña';
 export const EMAIL_ERROR = 'Por favor ingresar un email válido';
 export const PASSWORD_ERROR = 'Por favor ingresar una contraseña válida';
 export const WRONG_PASSWORD = 'Contraseña incorrecta';
