+++ src/app/screens/App/screens/Signup/strings.js
@@ -1,0 +1,1 @@
+export const SIGNUP = 'Signup';
