+++ src/app/screens/App/shared/strings.js
@@ -1,1 +1,3 @@
 export const BOOK_COVER_ALT_TEXT = 'Libro';
+export const EMAIL = 'Email';
+export const PASSWORD = 'Contraseña';
